+++ src/containers/CheckoutPage/CheckoutPage.js
@@ -15,6 +15,7 @@ import {
   isTransactionInitiateAmountTooLowError,
   isTransactionInitiateListingNotFoundError,
   isTransactionInitiateMissingStripeAccountError,
+  isTransactionInitiateBookingTimeNotAvailableError,
 } from '../../util/errors';
 import {
   AvatarMedium,
@@ -269,12 +270,22 @@ export class CheckoutPageComponent extends Component {
     );
 
     const isAmountTooLowError = isTransactionInitiateAmountTooLowError(initiateOrderError);
+    const isBookingTimeNotAvailableError = isTransactionInitiateBookingTimeNotAvailableError(
+      initiateOrderError
+    );
+
     let initiateOrderErrorMessage = null;
 
     if (!listingNotFound && isAmountTooLowError) {
       initiateOrderErrorMessage = (
         <p className={css.orderError}>
           <FormattedMessage id="CheckoutPage.initiateOrderAmountTooLow" />
+        </p>
+      );
+    } else if (!listingNotFound && isBookingTimeNotAvailableError) {
+      initiateOrderErrorMessage = (
+        <p className={css.orderError}>
+          <FormattedMessage id="CheckoutPage.bookingTimeNotAvailableMessage" />
         </p>
       );
     } else if (!listingNotFound && initiateOrderError) {
@@ -296,6 +307,12 @@ export class CheckoutPageComponent extends Component {
       speculateErrorMessage = (
         <p className={css.orderError}>
           <FormattedMessage id="CheckoutPage.providerStripeAccountMissingError" />
+        </p>
+      );
+    } else if (isTransactionInitiateBookingTimeNotAvailableError(speculateTransactionError)) {
+      speculateErrorMessage = (
+        <p className={css.orderError}>
+          <FormattedMessage id="CheckoutPage.bookingTimeNotAvailableMessage" />
         </p>
       );
     } else if (speculateTransactionError) {